Add files via upload
A program that creates the directed graph of the links between Wikipedia articles for a given depth of recursivity. 
We build a list of dictionary of nodes and a list of edges represented by couples of node
Each node is a dictionary associated with a specific wikipedia article:  {id, url, title, redirected, stats}
The keyword stats is still empty but is create for future applications of NLP methods.